#3. reward point 라벨 생성 및 구현

+++ BookStory/BookStory/Main/MainViewModel.swift
@@ -30,6 +30,6 @@ class MainViewModel {
     }
     
     func addRewardPoint(value: String, curPoint: Int, curLevel: Int) {
-        RewardConfig.addPoint(point: value.count * 2)
+        RewardConfig.addPoint(point: Float(value.count * 2))
     }
 }
